Start moving vds utilities
