+++ src/nexgen/tools/vds_tools/__init__.py
@@ -1,3 +1,13 @@
-from nexgen.tools.vds_tools.utils import VdsMapping, VdsSettings, find_datasets_in_file
+from nexgen.tools.vds_tools.utils import (
+    VdsMapping,
+    VdsSettings,
+    define_vds_dtype_from_bit_depth,
+    find_datasets_in_file,
+)
 
-__all__ = ["VdsMapping", "VdsSettings", "find_datasets_in_file"]
+__all__ = [
+    "VdsMapping",
+    "VdsSettings",
+    "find_datasets_in_file",
+    "define_vds_dtype_from_bit_depth",
+]
