+++ tests/tools/vds/test_utils.py
@@ -1,9 +1,13 @@
 import tempfile
 
 import h5py
+import numpy as np
 import pytest
 
-from nexgen.tools.vds_tools.utils import find_datasets_in_file
+from nexgen.tools.vds_tools.utils import (
+    define_vds_dtype_from_bit_depth,
+    find_datasets_in_file,
+)
 
 
 def test_find_datasets_in_file(nexus_file_with_single_dataset):
@@ -26,3 +30,12 @@ def test_find_datasets_fails_if_no_links_in_file():
     test_nexus_file["/entry/data/data_0001"] = [0, 0, 0]
     with pytest.raises(KeyError):
         find_datasets_in_file(test_nexus_file)
+
+
+@pytest.mark.parametrize(
+    "bit_depth, expected_dtype", [(8, np.uint8), (16, np.uint16), (32, np.uint32)]
+)
+def test_vds_dtype_from_input(bit_depth, expected_dtype):
+    d = define_vds_dtype_from_bit_depth(bit_depth)
+
+    assert d == expected_dtype
